+++ package.json
@@ -7,7 +7,8 @@
   "scripts": {
     "format": "bun --cwd app format && bun --cwd site format",
     "check": "bun --cwd app check && bun --cwd site check",
-    "test": "bun --cwd app vitest run && bun --cwd site vitest run"
+    "test": "bun --cwd app vitest run && bun --cwd site vitest run",
+    "build": "bun --cwd app build && bun --cwd site build"
   },
   "workspaces": [
     "app",
