+++ site/src/components/main_text/MultimediaAttachments.svelte
@@ -15,6 +15,7 @@
     media_type: string;
     link: string;
     name: string;
+    order: number;
   };
 
   type Props = {
