+++ site/src/components/main_text/TextRoutePage.astro
@@ -47,6 +47,7 @@ type Props = {
     media_type: string;
     link: string;
     name: string;
+    order: number;
   }[];
 };
