+++ site/src/pages/[key]/[...slug].astro
@@ -65,15 +65,15 @@ const child_items = resolved
   ? get_child_route_items(resolved.project_key, resolved.map, resolved.path_params, resolved.levels)
   : [];
 
-let media_links: any[] = [];
+let media_links: Awaited<ReturnType<typeof CACHE.media_links.get>> = [];
 if (resolved) {
   try {
     media_links = await CACHE.media_links.get(
       { project_id: resolved.project_id, path_params: resolved.path_params },
       cache_db_options_site
     );
-  } catch (e) {
-    // If project path doesn't exist or doesn't have media links, it's fine.
+  } catch {
+    // Path may not exist or have no media links.
   }
 }
 ---
